requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
faiss-cpu>=1.7.0
chromadb>=0.4.0
pypdf>=3.0.0
python-dotenv>=1.0.0
ollama>=0.4.0
flask>=3.0.0
requests>=2.28.0
werkzeug>=3.0.0
marked>=4.0.0
